backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
python-multipart
pydantic
langdetect
jieba
nltk
scikit-learn
wordcloud
pypdf
python-docx
python-pptx
markdown
Pillow
openai
graphviz
